+++ ChangeLog.md
@@ -43,6 +43,7 @@
 * Meilleur support du Tour de force
 * Meilleur support du Pacte Sanglant 
 * Permettre à un personnage d'être la Chair à canon de plusieurs autres
+* Rework Animation des Morts : meilleur support et automatisation
 
 ### Corrections de bugs
 * Correction des RD seulement contre perçant ou tranchant ou contondant.